+++ src/components/domain/index.ts
@@ -1,2 +1,3 @@
+export { default as TopNavigation } from "./TopNavigation";
 export { default as BottomNavigation } from "./BottomNavigtion";
 export { default as KakaoLogin } from "./KakaoLogin";
