+++ src/contexts/NavigationProvider/hook.ts
@@ -1,4 +1,6 @@
 import { useContext } from "react";
 import { Context } from "./context";
 
-export const useNavigationContext = () => useContext(Context);
+const useNavigationContext = () => useContext(Context);
+
+export default useNavigationContext;
