+++ src/contexts/NavigationProvider/index.tsx
@@ -48,3 +48,5 @@ const NavigationProvider = ({ children }: Props) => {
 };
 
 export default NavigationProvider;
+
+export { default as useNavigationContext } from "./hook";
