+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Button from "@components/Button";
-import { useNavigationContext } from "@contexts/NavigationProvider/hook";
+import { useNavigationContext } from "@contexts/NavigationProvider";
 
 const Home: NextPage = () => {
   const { useMountPage } = useNavigationContext();
